Organise payment components, tweak manual payment copy, tidy Spinner.
